pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fpdf>=1.7.2",
    "joblib>=1.5.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pybit>=5.11.0",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
    "xgboost>=3.0.4",
]
